+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Changed return type of `ConnectionConfig::with_encryption_trust_any_certificate() ` from `Result<Self, Error>` to `Self`.
  - Add support for Bolt 5.2, which adds notification filtering.
  - Add `Driver::is_encrypted()`.
+ - Reduce the number of lifetime generic parameters in `TransactionQueryBuilder` and `TransactionRecordStream`.
 
 ## 0.0.2
  - Update dependencies.